ecomerce/requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.1
djangorestframework==3.15.2
mysql-connector==2.2.9
mysql-connector-python==9.0.0
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.1
